+++ lib/gitcoin/getProjectDetails.ts
@@ -2,9 +2,9 @@ import type { Provider } from '@ethersproject/providers';
 import type { BigNumber } from 'ethers';
 
 import { getProjectRegistryContract } from 'lib/gitcoin/getProjectRegistryContract';
+import { getSafeOwners } from 'lib/gnosis/safe/getSafeOwners';
 import { fetchFileByHash, getIpfsFileUrl } from 'lib/ipfs/fetchFileByHash';
 import log from 'lib/log';
-import { getSafeOwners } from 'lib/safe/getSafeOwners';
 
 type MetadataOnchainDetails = {
   pointer: string;
@@ -75,8 +75,6 @@ async function getProjectOwners(ownerAddresses: string[], provider: Provider) {
 
     return acc;
   }, []);
-
-  log.info('🔥 safe owners:', owners);
 
   return owners;
 }
